\name{AnnotLists-package}
\alias{AnnotLists-package}
\alias{AnnotLists}
\docType{package}
\title{
Annotate multiple lists from a specific annotation file.
}
\description{
The lists of IDs must be in text tabulated format and placed alone in a folder.
The annotation file must be in text tabulated format and placed outside of the precedent folder.
The common identifiers must be placed in the first column in the lists and in the annotations file.
Batch processing of multiple lists is allowed.

Mor informations at http://www.mrbioinfo.com.
}
\details{
\tabular{ll}{
Package: \tab AnnotLists\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2011-10-21\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}

\author{
Nicolas Cagnard

Maintainer: Nicolas Cagnard <nicolas.cagnard@gmail.com>
}
\references{
http://www.mrbioinfo.com
}
\keyword{ package }
