\name{AnnotLists-package}
\alias{AnnotLists-package}
\alias{AnnotLists}
\docType{package}
\title{
Data extraction tool from annotations files.
}
\description{
Extract data annotations about lists of identifiers or microarray probes ID from an annotations file.
Batch processing of identifiers lists is allowed.

The annotations files and identifiers lists must be text tabulated files with one line of column's titles.

For multiple lists annotations, the lists must be placed alone in a common folder.

The first needed file is the Annotations file, then the identifiers file or the folder where are placed the files.

The identifiers must be in the first column of the lists and annotation files.

Mor informations about formats and uses at http://www.mrbioinfo.com.
}
\details{
\tabular{ll}{
Package: \tab AnnotLists\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-20\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}

\author{
Nicolas Cagnard

Maintainer: Nicolas Cagnard <nicolas.cagnard@gmail.com>
}
\references{
http://www.mrbioinfo.com
}
\keyword{ package }
